lxd/storage: Remove legacy zfs implementation
